Create Day 4: Class vs. Instance
These are Object-oriented concepts: Class and Instance
A Class is actually a blueprint or a representation of how an Object has to be instanciated. Whereas an Object is called as the Instance of a Class.
For example, Class can be assumed as a blueprint of a car, describing how it has to look, how it has to work, and so on. 
Whereas the Object is the real car.